替换autobancount
