--- webui/src/views/charts/components/2048.vue
@@ -1,19 +1,0 @@
-<!-- <template>
-  <a-card hoverable title="2048">
-    <div class="chart">
-      <vue2048 style="margin: 0 auto" />
-    </div>
-    <template #extra>
-      <a-tooltip :content="t('page.charts.tooltip.2048')">
-        <icon-question-circle />
-      </a-tooltip>
-    </template>
-  </a-card>
-</template> -->
-<!-- <script lang="ts" setup>
-import { Vue2048 } from 'vue-2048'
-import 'vue-2048/dist/style.css'
-import { useI18n } from 'vue-i18n'
-
-const { t } = useI18n()
-</script> -->
